Issue 112: Merged two feed fetcher implementations.
