--- app/src/test/java/org/apache/roller/planet/business/SubscriptionFunctionalTests.java
@@ -1,120 +1,0 @@
-/*
- * Copyright 2005 Sun Microsystems, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Source file modified from the original ASF source; all changes made
- * are also under Apache License.
- */
-
-package org.apache.roller.planet.business;
-
-import junit.framework.TestCase;
-import org.apache.roller.planet.pojos.Planet;
-import org.apache.roller.weblogger.TestUtils;
-import org.apache.roller.planet.pojos.Subscription;
-import org.apache.roller.weblogger.business.WebloggerFactory;
-
-
-/**
- * Test Subscription functionality.
- */
-public class SubscriptionFunctionalTests extends TestCase {
-    
-    private Planet testGroup1 = null;
-    private Planet testGroup2 = null;
-    private Subscription testSub1 = null;
-    private Subscription testSub2 = null;
-    
-    
-    protected void setUp() throws Exception {
-        // setup planet
-        TestUtils.setupWeblogger();
-
-        testGroup1 = TestUtils.setupGroup("subFuncTest1");
-        testGroup2 = TestUtils.setupGroup("subFuncTest2");
-        testSub1 = TestUtils.setupSubscription(testGroup1, "subFuncTest1");
-        testSub2 = TestUtils.setupSubscription(testGroup2, "subFuncTest2");
-    }
-    
-    
-    protected void tearDown() throws Exception {
-        TestUtils.teardownSubscription(testSub1.getId());
-        TestUtils.teardownSubscription(testSub2.getId());
-        TestUtils.teardownGroup(testGroup1.getId());
-        TestUtils.teardownGroup(testGroup2.getId());
-    }
-    
-    
-    public void testSubscriptionLookups() throws Exception {
-        
-        PlanetManager mgr = WebloggerFactory.getWeblogger().getPlanetManager();
-
-        // by id
-        Subscription sub = mgr.getSubscriptionById(testSub1.getId());
-        assertNotNull(sub);
-        assertEquals("subFuncTest1", sub.getFeedURL());
-        
-        // by feed url
-        sub = null;
-        sub = mgr.getSubscription(testGroup2, testSub2.getFeedURL());
-        assertNotNull(sub);
-        assertEquals("subFuncTest2", sub.getFeedURL());
-        
-        // count
-        assertEquals(2, mgr.getSubscriptionCount());
-    }
-    
-    
-    public void testSubscriptionGroupCRUD() throws Exception {
-        
-        PlanetManager planet = WebloggerFactory.getWeblogger().getPlanetManager();
-        
-        // retrieve subscriptions and add to group
-        Subscription sub1 = planet.getSubscriptionById(testSub1.getId());
-        Subscription sub2 = planet.getSubscriptionById(testSub2.getId());
-        Planet group = planet.getPlanetById(testGroup1.getId());
-        
-        // make sure no subs in group yet
-        assertEquals(0, group.getSubscriptions().size());
-        
-        // add
-        group.getSubscriptions().add(sub1);
-        sub1.setPlanet(group);
-
-        group.getSubscriptions().add(sub2);
-        sub2.setPlanet(group);
-        
-        planet.savePlanet(group);
-        TestUtils.endSession(true);
-        
-        // verify
-        group = planet.getPlanetById(testGroup1.getId());
-        sub1 = planet.getSubscriptionById(testSub1.getId());
-        sub2 = planet.getSubscriptionById(testSub2.getId());
-        assertEquals(2, group.getSubscriptions().size());
-        
-        // remove
-        group.getSubscriptions().remove(sub1);
-
-        group.getSubscriptions().remove(sub2);
-
-        planet.savePlanet(group);
-        TestUtils.endSession(true);
-        
-        // verify
-        group = planet.getPlanetById(testGroup1.getId());
-        assertEquals(0, group.getSubscriptions().size());
-    }
-    
-}
