+++ app/src/test/java/org/apache/roller/weblogger/business/PropertiesTest.java
@@ -14,6 +14,9 @@
  * limitations under the License.  For additional information regarding
  * copyright in this work, please see the NOTICE file in the top level
  * directory of this distribution.
+ *
+ * Source file modified from the original ASF source; all changes made
+ * are also under Apache License.
  */
 
 package org.apache.roller.weblogger.business;
@@ -55,7 +58,7 @@ public class PropertiesTest extends TestCase {
     }
     
     
-    public void testProperiesCRUD() throws Exception {
+    public void testPropertiesCRUD() throws Exception {
         
         // remember, the properties table is initialized during Roller startup
         PropertiesManager mgr = WebloggerFactory.getWeblogger().getPropertiesManager();
