--- app/src/test/java/org/apache/roller/weblogger/ui/UITestSuite.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- *  contributor license agreements.  The ASF licenses this file to You
- * under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.  For additional information regarding
- * copyright in this work, please see the NOTICE file in the top level
- * directory of this distribution.
- */
-
-package org.apache.roller.weblogger.ui;
-
-import junit.framework.Test;
-import junit.framework.TestSuite;
-import org.apache.roller.weblogger.planet.business.WebloggerRomeFeedFetcherTest;
-import org.apache.roller.weblogger.ui.rendering.util.CommentValidatorTest;
-
-
-/**
- * Test UI.
- */
-public class UITestSuite {
-    
-    public static Test suite() {
-
-        TestSuite suite = new TestSuite();
-	
-        // comment plugins
-        suite.addTestSuite(CommentValidatorTest.class);
-        
-        // custom planet fetcher
-        suite.addTestSuite(WebloggerRomeFeedFetcherTest.class);
-        
-        return suite;
-    }
-    
-}
